+++ c_sharp_for_home_assistant/CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.0] - 2021-07-05
+### Changed
+- Overhauled UI
 
+### Fixed
+- Manual run start for ReentyPolicy.QueueLatest and ReentyPolicy.Queue
+- Manual run stop for ReentyPolicy.QueueLatest and ReentyPolicy.Queue
 
 ## [0.5.0] - 2021-07-04
 ### Added